Fix all "conversion from ‘qsizetype’" compiler warnings
Variable, parameter and return types where changed to qsizetype when this
was trivial, required minor additional changs  or the type should logically be
qsizetype (e.g. for indizes).
In all other cases in assignments are wrapped in int(...), i.e. the qsizetype
result is cast to int.

+++ app/FileSystem/sizecalculator.cpp
@@ -173,5 +173,5 @@ void SizeCalculator::done()
 
 void SizeCalculator::emitProgress()
 {
-    emit progressChanged(((m_urls.length() - m_nextUrls.length()) * 100) / m_urls.length());
+    emit progressChanged(int(((m_urls.length() - m_nextUrls.length()) * 100) / m_urls.length()));
 }
